+++ routes/myAccountRouter.js
@@ -2,14 +2,28 @@ const { Router } = require("express");
 const {
   getManage,
   getMyAccount,
-  getActivate,
+  getActivateKey,
+  postActivateKey,
 } = require("../controllers/myAccountController");
 
 const myAccountRouter = new Router();
 
-// Get requests
+const isAuthenticated = (req, res, next) => {
+  console.log("isAuthenticated running...");
+  // If an account is logged in, redirect to the root
+  if (!req.isAuthenticated()) return res.status(401).render("401");
+  next();
+};
+
+myAccountRouter.use(isAuthenticated);
+
+// GET requests
+// User must be logged in to use these routes
 myAccountRouter.get("/", getMyAccount);
 myAccountRouter.get("/:username/manage", getManage);
-myAccountRouter.get("/:username/activate", getActivate);
+myAccountRouter.get("/:username/activate-key", getActivateKey);
+
+// POST requests
+myAccountRouter.post("/:username/activate-key", postActivateKey);
 
 module.exports = myAccountRouter;