+++ routes/postsRouter.js
@@ -1,0 +1,9 @@
+const { Router } = require("express");
+
+const postsRouter = new Router();
+
+// GET requests
+
+// POST requests
+
+module.exports = postsRouter;